requirements.txt
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
filelock==3.16.1
frontmatter==3.0.8
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
jiter==0.6.1
Markdown==3.7
MouseInfo==0.1.3
openai==1.52.2
packaging==24.1
pillow==11.0.0
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
PyYAML==5.1
requests==2.32.3
rubicon-objc==0.4.9
sniffio==1.3.1
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
